README: Apply via kubectl fix deprecation and improve behavior on failure.
There is some discussion going on about the kubectl run --serviceaccount flag. It is deprecated but still works: https://github.com/kubernetes/kubernetes/pull/99732
So for now, use the "overrides" workaround.

--restart=OnFailure - leaves the pod in status completed, if no errors occur. Otherwise it would restart and apply again, even on success.

Why don't we use a job instead of a pod?

It's likely that we would have to use YAML, because "kubectl create job" does not support --override. This makes handling it much more complicate for the user. For example, passing flags such as --remote, --argocd, etc.
Also interactive mode (""-i --tty") is not possible.
This makes it more complicate to wait for success and to view output. This would likely require a combination of "kubectl wait --for=condition=failure job/myjob" and "kubectl logs -f".

So for now, "kubectl run" seems to be the easiest way here.