Jenkins: Download plugins first, then install from files.
This makes sure to get deterministic version of plugins and all dependencies.

Removes waitForPluginInstallation() as this leads to failure because we install all plugins in arbitrary order. This order might lead to plugins being installed before their dependencies. As we do a restart at the end all plugins are loaded in proper order by Jenkins on startup.