k3d: Deactivate service-lb and traefik.
Right now no ingresses are used, so no need for an ingress controller, yet. No need for service-lb either, because we set up node ports for all applications. Also, the service-lb causes a lot of pending pods because lots of them try to bind to the same port.

This causes confusion with playground users. See:
https://community.cloudogu.com/t/gitops-playground-no-jobs-shown-in-jenkins/239/10