apply.sh paramters: Install fluxv2 only when --fluxv2 is passed
Historically, when passing no parameters, all GitOps operators were installed. With the switch to groovy, only Flux v2 is installed when no parameter is passed. This is a bug. But reflecting this, is might be less complicated to explicitly pass the parameter for the operators wanted. This is a breaking change in regard to the parameters, though.

On the other hand, we had this bug for quite a while. Now it is consistent. And it's also possible to not install a GitOps operator at all. Maybe the GitOps Playground is moving to provide general cluster infrastructure or even a platform 😱