Add ScmmRepo.writeFile
We want to move the responsibility for managing the temporary directory
from the caller to the ScmmRepo class.
This removes the burden to create the directory and ensure deletion from the caller.
We want to simplify the callers code regarding absolute paths.

As a result, we deprecate the original constructor to move away from
passing the temporary directory to the object instance.

Co-authored-by: Johannes Schnatterer <johannes.schnatterer@cloudogu.com>
Co-authored-by: Philipp Dziosa <philipp.dziosa@cloudogu.com>
Co-authored-by: Sentürk Ugras <sentuerk.ugras-extern@cloudogu.com>