Allow running apply.sh without container
This used to work unofficially, for development.
It stopped in 4d95f75e, when we (probably unintentionally) commented a line in init-jenkins.sh.
Later, in a9c1a160, we got rid of the need for a local groovy installation, which also influences the apply.sh, when run locally.

This commit updates the developers.md and fixes some small issues to allow local apply.sh execution for the sake of faster development speed.