Add template mechanism for yaml files
As we want to replace a lot of values in those files,
we think it is impractical to replace them each in code.
First, it is unintuitive to find where the change is made.
Secondly, we mix replacing yaml files and parsing yaml files in our code
base.

This approach using templating tries to solve these problems.
Furthermore, we think that it improves documentation as you immediately
see what would be changed in the templated file.