Use dependency injection
Using dependency injection allows us to be very explicit about
dependencies without needing to manually wire them with every change.
Explicitness helps us write tests, as it is obvious which dependencies
need to be mocked/changed/injected without risking to run code on a
real cluster.

Furthermore, by using @Singletons we gain the advantage to actually
implement services which use memoization.
For instance, the ScmmRepos don't need to be cloned as much, if we
utilize this.