Fix jenkins plugin installation
We were successfully uploading and installing the plugin
and then following the redirect to the progress page.
However, curl was removing the basic authentication.
As a result, we were encountering error 403 for the redirect,
not for the upload.

This is a bug in curl: https://github.com/curl/curl/issues/11486

Therefore, we temporarily replaced -L with --location-trusted for
jenkins