Fix bootstrapping config via dependency injection
We want to use dependency injection for the ApplicationConfigurator as
well.
However, we have transitive dependency on the Configuration object in
the chain, which at this point does not exist yet.
This dependency is - of course - not used in the exact path we are
using.
Therefore, we made the Configuration dependency optional using a
Provider<>.