Fix specifying protocol for ingress urls
In #147 we implemented ingresses for all applications.
The parameter --argocd-url was already used beforehand for urls in
notifications.
In #147 we mistakenly assumed that the url would only include the
hostname without protocol or paths.
Here, we fix this by parsing the URL and extracting the host for the
ingress configuration.