k3d: Fix ImagePullBackoff
for localhost images with --bind-localhost=true and registryPort != 30000.

The cluster tries to pull from e.g. localhost:32769 but the nodePort on the cluster was fixed to 30000.
Example:
docker pull localhost:32769/spring-petclinic-plain:202310051223-dc671a6-main
Works on host, but inside "k3d-gitops-playground-server-0" we get "connection refused", because the port there is 30000.

Confusing? Maybe the port-mapping of k3d-gitops-playground-server-0 helps:
docker ps
PORTS
0.0.0.0:32769->30000/tcp

The Solution is to adapt the registries' nodePort to the non-default value (e.g. 32769). This makes localhost:32769 also available for the kubelet in the server-0 container.
For this to work, the port range had to be adjusted to the maximum 65535.
Get rid of docker/values.yaml because we actually only need 2 params, that are easier to maintain within the code (bash for now, soon to be migrated to groovy)

Why did this work before?
Maybe it didn't work, but we never noticed, because --bind-localhost=true is mainly used on the CI-server, but there, we didn't test if the images are actually stated :/