Make welcome screen more static
When using the ingresses (#159) there is yet another dimension of logic we would have to maintain.
In bash, this kind of complexity just takes too much time to implement.
We might try building a new welcome screen once we fully migrated to groovy.

For the time being, this is less convenient for our users 🫤