init-cluster.sh: Restore default for bind-localhost
This was changed accidentally in #159 / 60c96aba because at this point we planned to switch to ingresses by default.
We later found out that subdomains of ingress don't work everywhere (e.g. not on debian).

So at this point it's easiest to keep downward compatibility. Also the docs and --help params state that bind-localhost is true by default.