Set PLAYGROUND_DIR in .sh files
Mainly because system property 'user.dir' seems to be empty in GraalVM.
But this way the
 .sh files are also more self-contained
and can be tested locally easier (if we ever need to that again).

The error encountered when running the prod image (GraalVM):

/tmp/workspace/ound_feature_call-groovy-earlier/scripts/utils.sh: line 90: java: command not found

Most likely because "$PLAYGROUND_DIR/apply-ng" was not found due to
PLAYGROUND_DIR being empty, because of System.getProperty('user.dir')
being empty.

Also get rid of apply-ng.sh, because it makes things more complicated
than
necessary.
It used to contain more logic, but this was moved to util.sh in 862201c2
.