Remove runGroovy calls from init-jenkins.sh
Fixes "utils.sh: line 90: java: command not found" caused by
[[ -f /tmp/workspace/ound_feature_call-groovy-earlier/apply-ng ]],
 because in Jenkinsfile image is run using docker.image.inside()
 which implictly sets docker run -w (--workdir) to Jenkins workspace.
 This leads to apply-ng binary not being found.

Migrating the calls to groovy avoids this problem.
As it also reduces technical debt it is a better invest than fixing the
problem  in bash.

Also move namePrefix logic from CLI to ApplicationConfigurator because
we also need this logic when using config file.