K8sClient: Properly escape params for secrets and CM.
Came up when testing email functionality with gmail:
Their app passwords contain spaces.
This resulted in the following error:
Stderr: error: no objects passed to apply

First idea was to just add quotes to the
 "--from-literal='${it.v1}=${it.v2}'"} string.
 This resulted in an error, that needed 4b7ca871 to be debugged.

Stderr: error: "'url" is not valid key name for a Secret a valid config key must consist of alphanumeric characters, '-', '_' or '.' (e.g. 'key.name',  or 'KEY_NAME',  or 'key-name', regex used for validation is '[-._a-zA-Z0-9]+')

or  Stderr: error: exactly one NAME is required, got 4.

So the only thing thats right is to do the tokenizing ourselves.
This way, the commands are inconvenient lists instead of strings.
But we escape the quoting hell.