Jenkinsfile: Send additional mail to devs if main fails
The merge commit of a PR at github has a git user with an email address
at users.noreply.github.com.

This leads to Jenkins mails on failed commits being not delivered.
So: send an additional mail to a list of devs defined in Jenkins.
A failed main branch deserves urgent attention.