Upgrade SCM-Manager to 3.1
Migrate from bash to groovy, while on it.

Why migrate?
System Property scm.initialPassword and scm.initialUser seem to no
longer work.
Instead. we can now use env vars via "extraEnv" value.
However, these are a string, not an object. This makes them very hard to
 set via helm CLI.
So instead of fiddling through the bash quoting hell, I decided to move
to a templated values.yaml which makes the whole process easier to read
and maintain.

While migrating to groovy, it seemed that the logic of scmm.url and
.host is more complex than necessary.
Also the "scmm-scm-manager.default.svc.cluster.local" is repeated in
multiple classes instead of being centralized in ApplicationConfigurator
.

Refactoring those would have been a good idea but was too much for the
current scope: Upgrading SCMM.

Migrating to groovy is a good first step.
Let's do the refactoring next time.

BTW: init-scmm.sh IS_LOCAL was unused and could be removed