Implement air-gapped mode for PrometheusStack
Helm does not have any options for changing dependencies. So our only
options is to alter the chart.

There, we would have two options for hosting the chart: Git or OCI.
Unfortunately it seems that using the OCI way in the local GOP is
impossible because we would have to use `helm pull --plain-http oci://.
.`, which not to be possible in Argo CD.
So we stick with what we already did with spring-boot-helm-chart: Use
git.