Prometheus: watch CRs in all namespaces
This allows application to create ServiceMonitors in their own namespace
.
Fits much better with our least privilege and multi-tenancy approaches.

On top of that remove the need to add a label to each ServiceMonitor.
This makes them easier to use.

Note that
any: true
instead of
matchLabels: {}
as described here:
https://github.com/prometheus-community/helm-charts/issues/106
fails, because not in schema.