init-cluster.sh: make ingress-port=80 default.
We introduced our own ingress controller and ingresses for all tools some time ago.

It since has become the most convenient option.

It's also the only option that works conveniently on Windows and Mac.

The docs always cite this option.

So why not make it the default, making GOP easier?

This removes the need to pass the option used almost always anyway.

Technically, this is a breaking change, though.
Before, by default, init-cluster.sh bound to localhost by default.
Now, it no longer does this.

Because this mode is less secure and linux-only and was indented for
development only, we don't think that this breaking change will affect
anyone in a negative way.