Fix startup error with --remote
GOP apply hang in an endless loop with "kubectl get svc scmm-manager
-n
default".
Which will not be available if not internal SCMM is present.

f15fe2b removed INTERNAL_SCMM from init-scmm.sh, but it was still used
in setExternalHostnameIfNecessary() (utils.sh).

There, the reference to the variable was hidden by eval.

Now we explicitly check INTERNAL_SCMM in init-scmm.sh, just like in
init-jenkins.sh.
This allows us to get rid of the evil eval statement.

Co-authored-by: Philipp <philipp.dziosa@cloudogu.com>