Prometheus: Move logic relating to values.yaml from groovy to YAML
Easier to maintain when all logic is in a single place.
Also less explicit.
Now that we have templating, let's make proper use of it!

Move DockerImageParser into template to work towards our goal of just
passing the whole config into templates to have the same keys in code
and templates.