Move createImagePullSecret() to trait
For easier reuse.

Using a trait instead of a util might also help us in the future.
For example, we can automate the setting of image values or extract all
images used.

An abstract class might have worked as well, but this way we can later
add more traits, if needed.