Move creation of example-apps pull secrets to new Content class
Registry does not have anything to do with this.
It's only for the content, because we add the secrets to the content
namespaces.

Also: The namespaces need to be created before adding the secrets.
 Why
should Registry.groovy create these Namespaces?
It's not a concern of the registry!

Similar to Mailhog, Vault, etc that create their own secrets.

The rest of the content remains in ArgoCD.groovy for now, but we will
refactor it to Content.groovy in the future.