Jenkins: Configures Prometheus only when necessary
This is the right thing to do in regard to GOP modularity.

It also reduces the risk of running into this notorious error:

java.lang.RuntimeException: Cannot enable authentication for prometheus: groovy.lang.MissingPropertyException: No such property: PrometheusConfiguration for class: Script1

Caused by running the Config Script on Jenkins that fails because the
Prometheus APIs are missing there.