README: Fix SVG diagrams not rendered properly in tab (#292)
When clicking on diagrams on GitHub they should be displayed in large.
However, on click the file is opened (still small).
When clicking on "Raw" the file opens full screen but is not rendered
properly.
Example:
https://raw.githubusercontent.com/cloudogu/gitops-playground/refs/heads/main/docs/gitops-playground-features.drawio.svg

Cause is Content-Security-Policy (CSP) used by raw.githubusercontent.com
:
img-src is limited, and data: URIs are not allowed.

To work around this, we could save the images as PNG.
PNG needs more size and is binary.
This would bloat our repo and slow down the loading time of our main
README.

Alternative: Add links that render the file via jsdeliver.
Disadvantage: Depending on a remote service.
It is only necessary for a quick full-screen view, though.
So, this solution seems pragmatic.

Minor fixes:

- Set height="100%" to SVG diagrams so opening them in full screen
scales them down to match the wide screen.
- Make clear the visualization tool is grafana
- Fix copyright background