Add build date to version number
This helps identify, if the running version is the intended one at
single glance.

E.g. untagged commit hashes that have been built a long time ago might
not be the intended one.
Or development versions with old build date might not be the intended
ones.

Example output:

With tag
1.0.0 (abc123, 2025-09-10 17:15)
or without tag
abc123, 2025-09-10 17:15